Update current frame move
A current frame move will move an entity that has movement during the current
frame.

+++ Assets/Client/src/Movements/Responses/Location.cs
@@ -5,10 +5,10 @@
     /// </summary>
     public sealed class Location
     {
-        public int X { get; set; }
+        public float X { get; set; }
 
-        public int Y { get; set; }
+        public float Y { get; set; }
 
-        public int Z { get; set; }
+        public float Z { get; set; }
     }
 }
